+++ examples/SimpleDualCall/SimpleDualCall/MainViewController.swift
@@ -1,0 +1,20 @@
+//
+//  MainViewController.swift
+//  SimpleDualCall
+//
+//  Created by Chance Kim on 2019/10/29.
+//  Copyright © 2019 remote monster. All rights reserved.
+//
+
+import UIKit
+
+class MainViewController : UIViewController {
+    
+    
+    override func viewDidLoad() {
+        super.viewDidLoad()
+        
+        self.navigationController?.setNavigationBarHidden(true, animated: false)
+        
+    }
+}